split up application file
