+++ flask_setup.py
@@ -1,0 +1,7 @@
+import os
+
+from dotenv import load_dotenv
+from flask import Flask
+
+
+application = Flask(__name__)
